+++ documentation/release_notes.md
@@ -16,6 +16,7 @@
 - matching-ml: requires a python environment with `Werkzeug<=2.2.3` because starting from `2.3.0` the webserver do not get the request headers 
 - [Wikidata tests are fixed](https://github.com/dwslab/melt/commit/ae824d9#diff-471d5bfa39673c940c5c7c2f7450ffe61545c052efd6ccf85080e0e589cbd7c9) because Wikidata entry for `EU` changed 
 - [Testcasevalidation is more memory friendly](https://github.com/dwslab/melt/commit/fe6915287637895c3fee63eec2e28934218ba0bd)
+- fixed macro F1 computation (it is **not** the harmonic mean of macro averaged precision and recall but now the macro averaged f1 scores)
 
 **New Tracks**
 - [pgx Track](http://oaei.ontologymatching.org/2023/pharmacogenomics/index.html)